Improve performance of rmSubmodule
Now, it's not necessary to 'git submodule deinit ...' since what we're looking for anyway is to completely remove any trace of the submodule. And to do that we just need to 'git rm SUBMODULENAME' and then 'rm -rf ./.git/modules/SUBMODULENAME'. After this you just need to commit the deletion.